Change Twitter authorize_url to oauth/authenticate
The authorize_url for Twitter was set to oauth/authorize, which caused a
reauthentication of the app each time a login was performed.  Changing it to
oauth/authenticate, as mentioned in
https://dev.twitter.com/docs/api/1/get/oauth/authenticate should ensure that
the login process is completed without requesting a new access token and the
user is automatically redirected back to the calling application.

Also changed the URL to use HTTPS, as specified by the Twitter
documentation.